+++ packages/cli/src/middlewares/index.ts
@@ -1,12 +1,4 @@
 export * from './auth';
 export * from './bodyParser';
 export * from './cors';
-import { filterListQueryMiddleware } from './filter.listQuery.middleware';
-import { selectListQueryMiddleware } from './select.listQuery.middleware';
-import { paginationListQueryMiddleware } from './pagination.listQuery.middleware';
-
-export const listQueryMiddleware = [
-	filterListQueryMiddleware,
-	selectListQueryMiddleware,
-	paginationListQueryMiddleware,
-];
+export * from './listQuery';
