+++ packages/cli/src/middlewares/listQuery/index.ts
@@ -1,0 +1,9 @@
+import { filterListQueryMiddleware } from './filter';
+import { selectListQueryMiddleware } from './select';
+import { paginationListQueryMiddleware } from './pagination';
+
+export const listQueryMiddleware = [
+	filterListQueryMiddleware,
+	selectListQueryMiddleware,
+	paginationListQueryMiddleware,
+];
